Create 8_puzzle_prblm_bfs
This file contains a Java implementation of a solver for the classic 8 Puzzle Problem. The 8 Puzzle is a sliding puzzle that consists of a 3x3 grid with eight numbered tiles and an empty space. The goal is to arrange the tiles in ascending order by sliding them into the empty space.